# first crack at extracting values from arrays by giving an index as a …