**Description**: What this PR does
goal is to extract values from arrays nested in deeper structures and values from within those array values

an example is provided in the added test

**Benchmark before change**:

``` go
docker run -v `pwd`:/go/src/github.com/buger/jsonparser -i -t jsonparser go test  -test.benchmem -bench JsonParser ./benchmark/  -benchtime 5s -v
testing: warning: no tests to run
PASS
BenchmarkJsonParserLarge-2                         50000            182450 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserMedium-2                       200000             33686 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserEachKeyManualMedium-2          500000             18086 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserSmall-2                       2000000              3361 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserEachKeyManualSmall-2          5000000              1905 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserEachKeyStructSmall-2          3000000              2776 ns/op             176 B/op          7 allocs/op
ok      github.com/buger/jsonparser/benchmark   59.980s

**Benchmark after change**:
docker run -v `pwd`:/go/src/github.com/buger/jsonparser -i -t jsonparser go test  -test.benchmem -bench JsonParser ./benchmark/  -benchtime 5s -v
testing: warning: no tests to run
PASS
BenchmarkJsonParserLarge-2                         50000            183814 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserMedium-2                       200000             34856 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserEachKeyManualMedium-2          500000             17943 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserSmall-2                       2000000              3403 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserEachKeyManualSmall-2          5000000              1877 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserEachKeyStructSmall-2          3000000              2786 ns/op             176 B/op          7 allocs/op
ok      github.com/buger/jsonparser/benchmark   60.314s
```