@buger 2016-08-20
Looks really cool!

Did you tried to re-use existing ArrayEach method, instead of embedding own parser?