@dparadise28 2016-08-20
Thanks! I'm glad you like it. I'm working on something that will be using the hard work put into your parser and I came across something i though would be useful so figured I'd see if i could contribute. 

Reusing ArrayEach was actually my first instinct and then i decided to go with somewhat duplicate parsing logic in an effort to be a bit more efficient and not iterate through part of the slice twice given you cant break when your using array each. But given the code is definitely more readable and more there is greater reuse in this approach I think your right on point. And maybe we can think of a way to break out of ArrayEach aft something is triggered in order to optimize the impl a bit further.