@rami-dabain 2016-08-26
for breaking from ArrayEach: https://github.com/buger/jsonparser/pull/58#issuecomment-242764996 though haven't tried it yet