@buger 2016-12-16
Nailed it https://github.com/buger/jsonparser/pull/77 !

Thank you for the first version! I had to completely rewrite it, and add support for EachKey.